ticker-research-mas/requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
anthropic>=0.40.0
yfinance>=0.2.40
duckduckgo-search>=6.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
tabulate>=0.9.0
